Refactor code to fit coding standards, defined in email sent
