+++ src/main/java/Command/Commands.java
@@ -1,4 +1,4 @@
-package Commands;
+package Command;
 import Models.Todo;
 import Duck.TaskList;
 import Duck.Storage;
@@ -18,7 +18,8 @@ public abstract class Commands {
      * abstract function to be implemented by all subclasses
      * @param list TaskList to be modified
      * @param storage Storage to be modified
-     * @throws IndexOutOfBoundsException is thrown when accessing a non existent object
+     * @throws IndexOutOfBoundsException is thrown when accessing a non-existent object
+     * @throws InvalidObjectClass thrown when function cannot be applied to object
      */
     abstract public void execute(TaskList<Todo> list, Storage storage, UI ui) throws IndexOutOfBoundsException, InvalidObjectClass;
