+++ src/main/java/Command/FindCommand.java
@@ -1,4 +1,4 @@
-package Commands;
+package Command;
 import Duck.Storage;
 import Duck.TaskList;
 import UI.UI;
@@ -27,15 +27,15 @@ public class FindCommand extends Commands{
      */
     @Override
     public void execute(TaskList<Todo> list, Storage storage, UI ui) throws IndexOutOfBoundsException {
-        boolean foundAny = false;
+        boolean isFound = false;
         String s = "";
         for (int i = 0; i < list.size(); i++) {
             if (list.get(i).getTitle().contains(search)) {
-                foundAny = true;
+                isFound = true;
                 s = s.concat(list.get(i).toString() + "\n");
             }
         }
-        if (foundAny) {
+        if (isFound) {
             ui.sendTextToUi("Here are the items:\n" + s);
             return ;
         }
