+++ src/test/java/Command/AddCommandTest.java
@@ -1,4 +1,4 @@
-package Commands;
+package Command;
 import Duck.Storage;
 import Duck.TaskList;
 import Models.Todo;
@@ -10,20 +10,14 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DeleteCommandTest {
+public class AddCommandTest {
     @Test
     public void executeTest() throws IOException {
         List<Todo> list = new ArrayList<>(); //fake list from file
         TaskList<Todo> taskList = new TaskList<>(list);
-        Storage s = new Storage("data", "duck.txt");
-
         Todo t = new Todo("title", false);
         Commands c = new AddCommand(t);
-        c.execute(taskList, s);
+        c.execute(taskList, new Storage("data", "duck.txt"));
         assertEquals(1, taskList.size());
-
-        c = new DeleteCommand(0);
-        c.execute(taskList, s);
-        assertEquals(0, taskList.size());
     }
 }
